etl_code/requirements.txt
metaflow==2.12.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
